Fix executing function in tracee.
Previously we called the function without adjusting the stack pointer.
There are two problems with this:
a) The 128 bytes below the stack pointer can be used for temporary
variables (without moving rsp). The function we interrupt and the
function we execute might interfer here.
b) rsp needs to be a multiple of 16 when calling a function (but not
during the execution of the function we interrupt).

Now we adjust rsp accordingly before executing the function in the
tracee. Note that all registers get restored before the regular
execution of the tracee continues - so rsp is restored as well.

Bug: http://b/182908805
Test: The previously flaky unit test works fine now.