Adjust Postioning System of Tracks - Make it Selfcontained.
CaptureViewElements have position and size information for their
appearence in the world coordinate system. This was only partially
used for tracks. The track's header/tab was not included by the
position/size of the track itself. Instead, the layout's gap size
between tracks contained the size of the header. The track relied
on this implicit space and drawed the header out of its scope.

This change adjusted the track's size. It is now the complete size
of a track.

Further, -- since there were bugs and it was unused -- this change
removes the handling for pinned tracks. The were drawn differently,
and remain untested.

Follow-up:
1. Simplify the Accessibility Interface (for tracks).
2. The nice thing about the state before this change, was that
   drawing the track's content did not required any knowledge about
   the header (and other margins). However, the clean solution would
   be to split the Track into a TrackHeader/Tab, (possibly multiple)
   TrackContentPane and TrackFooter.

Test: Take a capture including all types of tracks.