Set pid of FullGpuJob
`FullGpuJob` has a pid that `LinuxTracing` didn't set. But
`ProducerEventProcessor` copies that pid into `GpuJob` and
`CaptureEventProcessor` copies that pid into `TimerInfo`.

This was not a problem for the client as the pid is currently not shown.

As the frequency of these events is low and the impact on bandwidth is
negligible (the average number of bytes per event on a 10-second capture stays
around 25.0 in the range of measurement error), just set the pid.

I noticed this while adding tests for `GpuJob`s to
`LinuxTracingIntegrationTests`.

Bug: http://b/184942487

Test: Update unit tests. Capture Trata and see that gpu jobs in the timeline are
as usual.