Add MemoryTrack class
With this change, we add the MemoryTrack class such that the memory
tracks can be treated separately from the general GraphTracks.

We move the setting and drawing of `warning_threshold_`,
`value_upper_bound_` and `value_lower_bound_` to the MemoryTrack;
and meanwhile still leave the setting of `label_unit_` and
`value_decimal_digits_` to the GraphTrack.