Ground work for injecting liborbit.so
Instead of dynamically loading liborbit.so on the first api call,
Orbit will automatically inject the library when needed and patch
a global function table which will enable the API in the tracee.

This PR focuses on Api, OrbitBase and OrbitTest. Note that in its
current state, it breaks manual instrumentation in OrbitTest. This
is temporary as other PRs are in the pipeline. See PR #2140 for
end-to-end liborbit.so injection.