AccessibilityInterface now uses absolute positions for the rect.
Postions of CaptureViewElements as well as in AccessibilityAdapter::rect()
are absolute. Prior this change, we converted CaptureViewElement's postion
to a relative one in order to convert it back to an absolute one.
This makes the AccessibilityInterfaces unnessesarily complecated.

This change switch also to absolute positions in the accessibility.
Further, it introduces FakeTimerPane and FakeTrackTab as capture view
elements. Now all position calculations for accessibility are done
by AccessibleCaptureViewElement.

Follow up: The fake classes should be actual children of a Track.
  This requires the actual drawing to be also moved to those classes.

Test: Take a capture and inspect with the AccessibilityInsights.