Revert to std::map because they are more stable
Coming back the absl::btree_map to std::map because we have a crash that
could be related with stability and they don't guarantee that.

In an offline discussion, based on that we don't have a performance
requirement here, and we are having a flaky crash which is caused by a
multi-threading issue, we are coming back until the problem is properly
fixed.