Fix Tracing.Scopes test -- avoid missing events at the end.
With the fix of #2182 it can happen that introspection events
at the very end of a capture get lost. However, we've been a
bit restrictive here, as we also dropped scopes that have already
been scheduled on the thread pool, before this was shut down.

This can be improved (and in fact the Tracing.Scopes test relies on
this), by separating out an `IsShutdownInstantiated` out of `IsActive`.
This way, we only schedule scopes to the thread pool, when the thread
pool, has not yet started to be shut down, but scheduled scopes are
only aborted if the listener itself is not active anymore.

Further, this fixes the thread id check in the test itself for multiple
tests runs. The static variable was assigned only once, and not once
per test.

Test: Run test 1000 times.
Bug: http://b/185455477