Add a clang-format script to be invoked as pre-commit hook.
This is a first step into having an automated formatting upon commits.
It adds a script that can be executed from the pre-commit hook directly.

All the files touched (staged and unstaged) will be formatted completely
and the script will ask you which of the changes should be included
to the commit. Thus it also works with when working with hunks.

Test: Add "source path/to/file/" to the pre-commit hook script.