Add ORBIT_API_INSTANTIATE macro to be placed in user implementation
This is done to make the integration easier. If the api version changes,
all a user will have to do is update Orbit.h, not the integration code.
This also future-proofs the mechanism as we'll be able to modify the
code we want to have in the user implementation file.