Manage memory in tracee via unique_resource.
Added method AllocateInTraceeAsUniqueResource returning the memory
address wrapped in a unique_resource. Replaced almost all occurrences of
AllocateInTracee by AllocateInTraceeAsUniqueResource.
The only exception is AllocateMemoryForTrampolines: It is not obvious
yet if it is preferable to automatically free the memory or have this
done manually. Note that the memory for the trampolines stays allocated
after DetachAndContinueProcess.

Also remove implicit cast in UniqueResource.h.

Bug: http://b/185315205
Test: Unit tests.