Ask to Install "clang-format" "pre-commit" Hook on Bootstrab
This will ask the user that runs .bootstrap.sh if the new
clang-format pre-commit hook should be installed to the user's
hooks. It checks if the "pre-commit" script already exits, if not
it will be created. Further, it checks if the hook is already
installed. Only if it is not yet installed, the user will be asked.

Test: Run bootstrap.sh (tmp rename pre-commit file)