Inject liborbit.so using UserSpaceInstrumentation library
Orbit now takes care of loading liborbit.so into the target process. Orbit then overwrites a 
global function table in the target's address space, which activates the Api, given that the 
target has included Orbit.h and has placed "ORBIT_API_INSTANTIATE" in an implementation
file. This change makes it possible to activate the Api even if liborbit.so was not present on 
the system on the first Api call. This mechanism works for both C and C++.