Remove Qt::AA_UseDesktopOpenGL attribute
Yes this flag effectively makes Qt not fallback on OpenGL ES,
but it also disables Qt's OpenGL compatbility check. So we might
end up with OpenGL 1.1 instead of fallbacking back on Software
Rendering.

The correct solution is NOT to set this flag and not ship Qt with the
ANGLE OpenGL ES libraries.