Don't deploy ANGLE libraries
Since we don't want Qt to fall back on OpenGL ES we should not deploy
the ANGLE DLLs. This will effectively make Qt not consider ANGLE as a
fallback when proper OpenGL support is not offered by the graphics
driver. Qt will fall back on software rendering which is what we want.