Combine Gpu debug marker and submission tracks into one. (#2099)
Prior this, we had two Gpu tracks per queue (when the Vulkan layer
was enabled). However, those belong tightly together.

This change makes `GpuTrack` being a composed track of `GpuDebugMarkerTrack`
and `GpuSubmissionTrack`. It refactors out the logic of each kind of
prior GpuTrack.

In order to do so, tracks can be now transparent and may have an intent.

Test: Take a capture with the Vulkan layer.
Bug: http://b/182751932.