Add GetTracerPidOfProcess
GetTracerPidOfProcess returns the pid of the attached tracer process
for the given pid. If no process is attached, the function returns 0.
Internally, the function reads the content of "/proc/<pid>/status" to
find the "TracerPid" entry and returns its value.

Tested with the current process with and without a debugger attached.