Add GetTracerPidOfProcess to detect existing "Tracer" process
* Add GetTracerPidOfProcess

GetTracerPidOfProcess returns the pid of the attached tracer process
for the given pid. If no process is attached, the function returns 0.
Internally, the function reads the content of "/proc/<pid>/status" to
find the "TracerPid" entry and returns its value.

Tested with the current process with and without a debugger attached.

* Check that target has no tracer process before attaching

We rely on ptrace for our userspace instrumentation. There can only
be a single tracer process attached at a time. We need to detect
any tracer and, if found,  warn the user that Orbit won't be able
to perform any userspace instrumentation.