Rename orbit_code_viewer::Dialog::SetSourceCode
This function name is confusing because its also the setter in the
assembly case. It even gets more confusing when we introduce the
combined view.

So this commit renames it to `SetMainContent` which I don't like
but I don't have a better suggestion. In a subsequent commit I will also
introduce annotating content which refers to lines in between the main
content.