Add line number types to orbit_code_viewer::Viewer
This change allows each line in a Viewer to also hold some metadata. For
now it allows to set a custom line number value and a content type.
Either it's classified as main content or annotating content and is
meant to fulfill the use-case of showing both assembly and source code.