Patch Callstacks of Leaf Function Samples.
In leaf functions (or samples that hit the prologue/epilogue),
`$rbp` might not point to the beginning of the frame. In those cases
perf_event_open will miss the sample's direct caller in the resulting
callstack (on frame-pointer-based unwinding).

This change will in addition to perf_event_open, unwind the first
two frames with libunwindstack, and if the sample's caller is
different, we know the missing pc. We will patch the sample in this
case.

Now that we handle the case that our sample falls inside the prologue,
the test does not need to discard those samples anymore.

Test: Compile a target with -momit-leaf-frame-pointer, inspect the
      callstack. Further, run the integration and unit tests.

Bug: http://b/160399871