Remove special handling for outer [uprobes] in LibunwindstackUnwinder
This was introduced because for a while we were reconstructing callstacks
falling inside a dynamically instrumented function or one of its descendants
by collecting samples on uprobes and concatenating them ("callstack stitching").
This went away a long time ago with patching the stack with the return address
collected on uprobes ("stack patching"), so let's remove this special case,
it's confusing.

Test: Run LinuxTracingIntegrationTests a few times. Capture Trata while
instrumenting functions and verify callstacks are in order.