Add `CallstackType type` to `orbit_client_protos::CallstackInfo`
Note that the logic of
`CallstackData::FilterCallstackEventsBasedOnMajorityStart` had to be slightly
changed (the method was renamed accordingly): instead of throwing away
`CallstackEvent`s, we now change the type of the referenced `CallstackInfo`.

Bug: http://b/188178748
Bug: http://b/188498206

Test: Update affected unit tests. Capture Trata and verify callstacks (the ones
that are not unwinding errors) are in order in the timeline and in the
aggregated views. Play around with selections to verify that non-`kComplete`
callstacks don't end up in them.