Correct function order in UpdateTrackOrder
Regression from 1.63 (more precisely from https://github.com/google/orbit/pull/2221).

In that PR we alter the order by mistake of the functions "UpdateMovingTrackSorting" and "UpdateVisibleTrackList".

This PR we are reverting that change and renaming UpdateMovingTrackSorting to a name which explicit the fact that we are updating the position in the visible tracks.

It solved http://b/188521605.