Introduce a syntax highlighter to annotating lines
Make HighlightAnnotatingBlock which highlight every character using the same color.
To see if a line is MainContent or AnnotatingLines, we are also introducing HighlightingMetadata,
an abstract base class in SyntaxHighlighter which provide a method to see if a line is MainContent.

With this, all the details from Metadata won't be visible to SyntaxHighlighter.