Always hook vkQueuePresentKHR in FakeClient to output avg frame time
But only if the target has loaded `libvulkan.so.1`.

This might slightly affect the existing metrics in `OrbitService` performance
testing, but it seems only within measurement error. See below.

Bug: http://b/187827094

Test:

Tried on Trata, Infiltrator, and a binary that doesn't to graphics.
With Trata and Infiltrator, compare results with Orbit UI's frame tracks and
with the "Frame rate rendered" of the Stadia extension.
Note that Infiltrator doesn't hit 60 fps.

Compared the results of the OrbitService performance testing on YHITI with the
previous and with this version of `OrbitFakeClient`.
Previous: http://sponge2/6cd37b5c-792f-47b1-84db-6c4240377e77
This change: http://sponge2/8114a173-2bc6-4e7f-a165-1f6658eb3150
There doesn't seem to be any relevant difference.