Add avx2 and fma flags to linux profiles.
I need avx support for a test (checking proper handling of m256
parameters in the context of user space instrumentation). This
could be handled in a way that only the test is affected and strictly
speaking I only need avx.
But I guess there should be pretty little reason not to add avx2
support globally.

Test: Compiled / ran unit tests with clang9 and gcc8. Let's see if the
CI works finds something.