Remove TimeGraph dependency from LiveFunctionsDataView
All data views will be moved into its own module for better testability.
For this to happen they need to be decoupled from TimeGraph and their
pointer to `OrbitApp` needs to be replaced by a pointer to an abstract
base class of `OrbitApp` (probably called `AppInterface`).

This is just a first step and only affects LiveFunctionsDataView.