Don't send callstacks with no frames at all, they're kind of meaningless
For the unexpected cases where we really have no frame: do nothing instead of
sending an unwinding error. Again, these cases are unexpected and we haven't
observed them, but we don't want to just crash.
For the samples falling inside uprobes: include the frame we have, even if it's
inside uprobes code.

Test: Take some captures with Trata, with and without frame pointers, with and
without dynamic instrumentation of a very frequent function. Check for crashes.