Add "Unwind errors" column to "Sampling" tab
The delicate part is updating `PostProcessedSamplingData` and
`SamplingDataPostProcessor`.

Note that the top-down and bottom-up views still don't show that information,
but instead they simply not add the non-`kComplete` callstacks. This means that
the number of samples in the "Sampling" tab is actually different from the
number of samples in the top-down and bottom-up views. Worry not, this is all
tackled in the next commit.

Possibly distinguishing between different error types will come at a later
time.

Bug: http://b/188496245
Bug: http://b/189824181

Test: Update `SamplingDataPostProcessorTest`.
Capture Trata a few times, waiting for unwinding errors. A way to get a high
percentage of them is to instrument a function that is called very frequently.
Verify numbers make sense and callstacks are in order, including the ones
corresponding to errors.