Relocate instructions.
Tooling to move instructions in memory. Interesting instructions are
e.g. relative jumps or instructions using ip relative addressing.

In order to keep the size of the PR limited this contains only the
code to move one individual instruction so there are no tests yet
that actually run the relocated instruction. This will follow in the
next PR.

Bug: http://b/183113989
Test: Unit tests in PR