Add [unwind errors] statistics to top-down and bottom-up views
This is how the top-down tree looks now when there are unwinding errors in a
thread (or in the "(all threads)" tree):
```
- thread [1234]
  - clone
    + start_thread
  - [unwind errors]
      foo
```

Here is an example for the bottom-up view when there are unwinding errors for a
function:
```
- foo
  - bar
    + main
  - [unwind errors]
      thread [1234]
      other_thread [1235]
```

I made the "[unwind errors]" string red (only that, not the whole row). I don't
think it's distracting, it doesn't appear often.

Possibly breaking down the statistics by error type will come at a later time.

Bug: http://188496245
Bug: http://189823771
Bug: http://189824607

Test: Played around with it quite a bit with Trata and Infiltrator, with and
without instrumenting a function called very frequently. Tried selections,
context menus, searches, copies.