Make ThreadPool inherit from Executor
This will allow ThreadPool to use `ScheduleAfter` and
`ScheduleAfterIfSuccess` which hasn't been available previously.

It also requires ThreadPool to always be instantiated as shared_ptr since
Executor uses `weak_from_this` to do life time management in
`ScheduleAfter` and `ScheduleAfterIfSuccess`.

While this is an additional overhead for ThreadPool, it is negligible
because we neither have many instances of ThreadPool, nor we are passing
these instances around.