Add SimpleExecutor
This executor is meant to be used in tests. It's a simple replacement
for MainThreadExecutor and it doesn't require Qt. Since ThreadPool now
also inherits from Executor it can also be used for tests which
previously required ThreadPool.

It might also be interesting for folks who want to use promises and
futures in parts of Orbit which can't depend on Qt (like the service).