Improve OrbitBase post build commands
Beforehand there were two commands, one for cleaning up old testdata and
one for copying it. These commands are not guaranteed to be called one
after the other, but could also be called in parallel. Now its one
add_custom_command with 2 COMMAND actions in it, which makes sure they
are executed one after the other.