Add Flag to Configure Stack Dump Size in Sample.
This makes the size of the stack dump to copy by perf_event_open
be configured using the "--stack_dump_size=" flag. Both for
DWARF-based unwinding, and for frame-pointer-based unwinding with
leaf functions.

The default value is the 65000 bytes from our current DWARF-based
handling. For frame-pointers, we should set this way lower. That
will be easier when moving to an UI option.

Test: Take a capture, Run LinuxTracingIntegrationTests
Bug: http://b/190491042