Make DataView independent from OrbitApp
The required pointer to OrbitApp is replaced by a pointer to
AppInterface.

Since all the subclasses still require a pointer to the "real" OrbitApp
(AppInterface does not yet declare all the necessary member functions),
temporarily a shadowing member variable with the same name `app_` gets
introduced. This will be removed again in subsequent commits.

By using the same name `app_` I avoid changing a lot of code from `app_`
to a different name and back again later on.