Move gtest's main functions into separate module
gtest's main functions used to be created by CMake on the fly. This
commit is moving them into its own cmake module and subdirectory and
treats their targets as if they were standard library targets.

It is also changing when QApplication and QCoreApplication instances are
created. By using a test listener we postpone the creation to right
before the test start.

That's important because this way gtest command line options like
`--help` or `--gtest_list_tests` won't need a QApplication instance and
hance they work in headless setups.