Remove intersecting modules on AddOrUpdateModuleInfo
Also fix a bug where module was incorrectly found when looking
for it by the end address. Note that end address could be
another module start address.

Prior this change following use-case behaved incorrectly:
Two concecutive modules (a and b) are unloaded and 1 bigger
module (c) is loaded in the same address range. In this case
old logic ends up returning incorrect module for
addresses >= start address of module (b).

Note also that this change fixed the state of the snapshot
it does not provide any additional information about
timeline, in other words you still can not request a state of
module map at specific timestamp. If the order of events is
preserved this snapshot can be used to process other events
as we go.

Test: run ClientDataTests
Bug: http://b/190805289