Collect RssAnon instead of Rss for the process
With this change, we collect the size of resident anonymous memory `RssAnon`
rather than the resident set size `Rss` for the process memory usage.

`Rss` = `RssAnon` + `RssFile` + `RssShmem` and `RssFile` counts the shared files.
However the cgroup memory usage counts a mapped file only if the cgroup is the
owner. Using `Rss` as the process memory usage will result in the case
that process memory usage is larger than the cgroup memory usage.