Simplify TriangleToggle State and Usage. (#2455)
Before this change, the track's triangle toggle had a tri-state,
expanded, collapsed, inactive (meanning not-collapsible). This
required the maintainance of the "initial state", but most
importantly, IsExpanded() was not !IsCollapsed(), which is
error-prone.

Now, similar to the track itself, the toggle can be collapsible
(or not) and independent of this, the toggle has is collapsed
(or not). The default is not collapsed, and (but) collapsible.

Test: Take a capture with all kinds of tracks. Collapse/uncollapse.
Bug: http://b/185465747