Add LoadBias and GetExecutableSegmentOffset to PE/COFF object files
Symbols addresses stored in DWARF debug info in COFF files have a
preferred load address added to them (which is not used anymore). We
need to subtract this address in the same way as the load bias for ELF
files. We also add the executable segment offset ("base of code" in
PE/COFF) for PE/COFF files and move up the common methods to this and
the load bias into the ObjectFile class.