Fix death tests in Qt-based modules on Windows
Death tests have been broken in Qt-based modules. These modules require
a QApplication object to be constructed before they are executed. That
used to be done by a Gtest event listener which does not receive any
events in case of a death test.

So this commit is adding an additional work around for death tests.