Improve prefix of function/thread name when copying from CallTreeWidget
For the leading spaces, use no-break spaces. Also change the characters that
represent the expanded and collapsed states.
Without this, pasting to a spreadsheet causes leading spaces to be removed, and
the strings starting with the plus sign (still the case of top-level nodes) to
be considered formulas, resulting in errors.
The new characters also match the source of the copy better.

Bug: http://b/177954828
Bug: http://b/191365691

Test: Copy from `CallTreeWidget` into Google Sheets and Excel online.