Only copy percentage from columns that contain two values
This way the values of these columns can be interpreted as numbers by
spreadsheets.
Of course this change is not ideal since it drops some information when copying
(the raw numbers). But having a different number of columns to copy than are
displayed is not straightforward for either `DataView`s and for the
`CallTreeWidget`, so I'm avoiding that.

Bug: http://b/191365691

Test: Copy from `DataView`s and `CallTreeWidget`s into Google Sheets.