Remove "intermediate" MetadataEvent proto
Move contained events up, to `ProducerCaptureEvent` and `ClientCaptureEvent`.
Also removed unused `InfoEvent`.
This is mostly moving some code around, adding some trivial methods, and
adding tests for those methods (all similar to each other).

Note: We don't need to take care of backwards compatibility because
`MetadataEvent` was introduced *after* 1.65, so we are still in time to change
it before the 1.66 branch cut.

Bug: http://b/192857660

Test: Take some captures in conditions that generate some of the events
involved.