Introduce "View" menu
This moves view- and windows-related settings into the new "View" menu.
To keep the UI consistent, it also changes the hotkeys 'X' and 'H' to
be QT-owned global hotkeys (Ctrl + X and Ctrl + H) that update the
state of the associated action buttons.

'X' is still supported if the capture window is in focus, but is no
longer displayed in the help text.