Add tooltips for SystemMemoryTrack
With this change, we add tooltips for the track tab and legends of
SystemMemoryTrack. When the user move the mouse over the legend, a
tooltip box pops up and shows the meaning of this data and how it
is collected.

Besides, we create the SystemMemoryTrack class and move the initialization
parts (e.g., track name, value label unit, series names) from TimeGraph and
TrackManager to the track constructor.