Add tooltips for CGroupAndProcessMemoryTrack
With this change, we add tooltips for the track tab and legends of
CGroupAndProcessMemoryTrack. When the user move the mouse over the
legend, a tooltip box pops up and shows the meaning of this data and
how it is collected.

We create the CGroupAndProcessMemoryTrack class and move the
initialization parts (e.g., track name, value label unit, series names)
from TimeGraph and TrackManager to the track constructor.