Add tooltips for PagefaultTrack
With this change, we add tooltips for the track tab and the legends of
PagefaultTrack. When the user move the mouse over the legend, a tooltip
box pops up and shows the meaning of this data and how it is collected.

Also, we create the PagefaultTrack class and move the track initialization
parts from TimeGraph and TrackManager to the track constructor.