Proper handling of function_id in the trampolines.
Previously trampolines handed over the function address to the entry
payload. Since Orbit expects the functions to be identified by a
the function_id this PR changes the name of this parameter to
function_id.
There is an additional complication however: The function_id is not
stable accross multiple profiling runs. Therefore the trampolines
need to get patched with the current function_id before each run. This
is done in `InstrumentFunction`. The offset at which the function_id
needs to be inserted is a compile time constant (gets verified during
the construction of the trampoline code).