Rework on the graph track line drawing
We had the issue that some horizontal lines in the graph track are
missing when we are drawn at the margin of pixels.

This was fixed previously by increasing the line width from 1px to 2px.
However, this affects the other tracks, which is not expected.
We change to draw lines at the center of pixels. This suppose to fix the
missing line issue in most cases.