Fix generate_coverage_report.sh css linking
Beforehand the css used for the main index.html page was simply the css
file of the last generated module report. Unfortunately the last module
is currently VulkanTutorial, which does not contain unit test and
therefore no report and no css file is generated.
This commit changes the used css file to always be the one from the
module OrbitQt.