Right align annotation with graph content
The previous computation of annotation right x utilizes
`TimeGraphLayout::GetSliderWidth()`, which means it always assumes the
existence of vertical slider. When the vertical slider does not show up
in the capture window, the annotation and the graph content are not
right aligned.

We fix this by computing the track right margin by
`TimeGraph::GetRightMargin()`, rather than
`TimeGraphLayout::GetRightMargin()` +
`TimeGraphLayout::GetSliderWidth()`.