Fix compiler flags in cross build
Both the ggp_sdk package - which contains the compiler when compiling
for Stadia - and our profiles define C and C++ compiler flags. Both
lists of flags need to get merged to properly compile Orbit, but this is
currently not happening in all cases. Conan recipes that don't use CMake
(i.e. openssl) won't use the flags from the ggp_sdk package since they
get overwritten by the flags defined in the profile.

As a workaround we pass our compiler flags from the profile as a package
option to the ggp_sdk package. The package will merge the flags and set
them properly.