Replace TextBox for TimerInfo in the UI
In this commit we are replacing in the UI all the references to TextBox
for TimerInfo as basically TextBox is now a container who almost only
hold TimerInfo.

We are also replacing the only method left in the TextBox (Duration) for
timer_info.end()-timer_info.start().

And finally, we are replacing a few pointers for const references when
we are assuming than the pointer shouldn't be nullptr.

Test: Starting/Loading a capture. Everything seems normal, no
performance regression here.