Fix flaky memory integration test
In memory integration test, we check whether memory events (e.g.,
`SystemMemoryUsage`, `CGroupMemoryUsage` and `ProcessMemoryUsage`)
collected in the same sampling window are sampled at a very close time.
This is done by comparing the difference of memory event timestamps to a
tolerance value `kMemoryEventsTimeDifferenceTolerace`.

The previous tolerance value is too small and so results in some flaky
memory unit tests. We increase the value with this change.