Add LLVM and clang to ggp docker container
The llvm-core recipe does not support proper cross-compilation and
expects an installed version of llvm-tblgen and clang-tblgen on the host
machine. We can work around that problem by installing the debian
packages.

For clarification: This is not a new limitation. It has been there
before with LLVM-9 as well. But the docker containers have recently been
simplified and some packages had been removed which made it impossible
to recompile the LLVM packages from source. This is fixing it now.