Move indentation_level to Draw methods
As it is only used to determine some sizes inside the Draw method, and
it isn't some track inherent property we move it there. This allow us
to erase indentation_level_ from Tracks.

In the future, if we are interested in allowing tracks to be moved in
top of another, this change could help in simplify the logic of that
change.

Also, we move the size of the collapse track button to TimeGraphLayout.