Further increase IDLE_TIME_ON_EMPTY_{RING_BUFFERS,DEFERRED_EVENTS}_US
We have seen in go/stadia-orbit-service-vs-perf that we still spend quite some
time calling `usleep`, so let's increase this further as the sizes of the
`perf_event_open` ring buffers allow for much longer times without being polled
(necessary if `OrbitService` is not scheduled for a while).

When capturing nothing with
```
./OrbitFakeClient --pid=10189 --frame_time=false --gpu_jobs=false \
--scheduling=false --sampling_rate=0
```
CPU utilization of OrbitService goes down from ~1.5% to ~0.5%.

This benefit is even higher on normal captures as the number of context switches
that end up being recorded is lower, as `OrbitService` itself causes fewer.

Test: Comparing with main when instrumenting Trata's `execute_task`: the
frequency of the (rare) lost/discarded events doesn't seem to increase.