Represent chunks of memory in the tracee by a class.
Previously we only handed out pointers (in the form of uint64_t's).
Going forward we will also need the size and some book keeping
information since we'll need to adjust the access protection while
the program is running. The memory will get allocated with write
permissions and will be made executable (and non-writable) when
needed.

This change is more or less a no-op in such that it merely replaces
the old system with the new class without using any new functionality.