Solve issue with OpenGl lines
We had an issue with some lines which weren't visible in GraphTracks
(http://b/192195534). The solution was move them to the center of the
pixel.

This change is making a new issue because they aren't aligned with
the boxes (http://b/194681684).

In this PR, we are reverting the previous solution and using an OpenGL
directive to force lines to be as smooth as possible but always
visible.

Test: Start a capture. Check graph horizontal lines are always visible.
Check Boxes are at the same y-level than vertical lines.