Add loading of symbols from PDB files (not integrated)
We add a PdbFile class that allows loading symbols from a given PDB
file. This is not yet integrated into the symbol loading flow.

For testing, this adds a basic .dll and .pdb compiled from dllmain.cpp.
The latter is also added in this commit together with instructions how
to build the .dll and .pdb.

Tested: Unit tests added.
Bug: http://b/194497774