Send new manual instrumentation events.
This changes the existing implementation of manual instrumentation
to send the new protos. Note, that this does not yet require the
version to be bumped up, as it does not change the header.

Some words about performance:
We measured the overhead of manual instrumentation induced to the
target per scope (start+stop). Before this change we induced about
235ns and after this change we are about 290ns.

The average message size send to the client dropped from 31 bytes
to 20 bytes.

Also the CPU utilization of the PSSI::RcvEvents thread (in game)
decreased by 10 percent points, while the SenderThread increased
by 5 percent points. The ForwarderThread remains at the same level.

Test: Run OrbitTest
Bug: http://b/194763537; http://b/194764373; http://b/194764259;