Fix regression with horizontal zoom
With the change of not including the Timegraph's right margin in the
time range, but including them to calculate the mouse_ratio we are
having an inconsistency every time we zoom. This is a big issue because
zoom-in/out is not an inverse operation (http://b/195652915).

In this change, we are including it and also clamped the max and min
ratio for the zoom triggered in the margin.