Use llvm-cov-10 in code-cover-report presubmit check
A recent PR triggered a bug in llvm-cov-9 which refuses to read coverage
data from a static library.

The bug seems to be fixed in LLVM-10. At least it works fine with
llvm-cov-10.

So this commit updates the corresponding docker container to provide
llvm-10 in addition to clang-9. (We will still use clang-9 for
compilation to stay aligned to other presubmit checks.)